+++ webapp/tests/integration/admin-logs-view.test.js
@@ -3,8 +3,8 @@
  * @tagline         Integration tests for admin logs view functionality
  * @description     Tests for admin logs page, docTypes integration, search functionality, and error handling
  * @file            webapp/tests/integration/admin-logs-view.test.js
- * @version         1.4.6
- * @release         2026-01-06
+ * @version         1.4.7
+ * @release         2026-01-07
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/