+++ webapp/tests/unit/controller/handlebar-variables.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for W-103: {{let}} and {{#with}} for custom variables
  * @description     Tests for variable assignment and context switching functionality
  * @file            webapp/tests/unit/controller/handlebar-variables.test.js
- * @version         1.4.6
- * @release         2026-01-06
+ * @version         1.4.7
+ * @release         2026-01-07
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/